Split PRESEEA class into mixins
