+++ preseeapy/PRESEEA.py
@@ -2,7 +2,6 @@ from .CorpusDefinition import Corpus
 from twisted.internet import reactor
 import json
 import csv
-import itertools as it
 import os
 from multiprocessing import Queue
 from scrapy.signalmanager import dispatcher
@@ -10,9 +9,14 @@ from scrapy.crawler import CrawlerRunner
 from scrapy import signals
 from .preseeaspider.spiders.preseeabot import PreseeabotSpider
 from .utils import ProcessHandler
+from .CityCorpusMixin import CityCorpusMixin
+from .AgeCorpusMixin import AgeCorpusMixin
+from .GenderCorpusMixin import GenderCorpusMixin
+from .EducationCorpusMixin import EducationCorpusMixin
 
 
-class PRESEEA(Corpus):
+class PRESEEA(Corpus, CityCorpusMixin, AgeCorpusMixin,
+              GenderCorpusMixin, EducationCorpusMixin):
     """This class describes the PRESEEA corpus and enables
        scraping its webpage: https://preseea.linguas.net/Corpus.aspx
 
@@ -34,10 +38,18 @@ class PRESEEA(Corpus):
             self._feature_dict = json.load(file)
 
         # PRESEEA specifc filter values
-        self._city = ""
-        self._gender = ""
-        self._age = ""
-        self._education = ""
+        self.set_city("")
+        self.set_gender("")
+        self.set_age("")
+        self.set_education("")
+
+    def __str__(self):
+        return "PRESEEA" + "_"\
+            + CityCorpusMixin.__str__(self) + "_"\
+            + GenderCorpusMixin.__str__(self) + "_"\
+            + AgeCorpusMixin.__str__(self) + "_"\
+            + EducationCorpusMixin.__str__(self) + "_"\
+            + "Phrase:" + self.get_search_phrase()
 
     def retrieve_phrase_data(self) -> list:
         """Retrieve phrase data with a separate process.
@@ -56,41 +68,6 @@ class PRESEEA(Corpus):
         process_instance.close()
 
         return phrase_list
-
-    def retrieve_city_info(self) -> int:
-        """Get general information about a city.
-
-        Args:
-            n_total (int): Total number of samples for a city
-        """
-        city_list = list(self._feature_dict['City'].keys())
-        if self._city not in city_list:
-            Warning('City not available in Corpus.')
-            return None
-
-        # Number of samples per city
-        n_total = self.get_number_city_samples()
-
-        return n_total
-
-    def get_number_city_samples(self) -> int:
-        """Get number of available samples for a city.
-
-        Returns:
-            n_total (int): Total number of samples for a city
-        """
-        # Temporarily save search phrase
-        temp_search_phrase = self._search_phrase
-        # Use empty search phrase to get available samples
-        self._search_phrase = " "
-
-        # Get number of samples for that phrase
-        n_total = self.get_number_samples()
-
-        # Set back correct search phrase
-        self._search_phrase = temp_search_phrase
-
-        return n_total
 
     def get_number_samples(self) -> int:
         """Get number of samples for a specific phrase.
@@ -549,78 +526,13 @@ class PRESEEA(Corpus):
 
         return data
 
-    def set_city(self, name: str):
-        """Set Corpus' instance city by its name and check beforehand
-
-        Args:
-            name (str): Name of the city the corpus should be filtered on
-        """
-        city_list = list(self._feature_dict['City'].keys())
-        if name in city_list:
-            self._city = name
-        elif name == "all":
-            self._city = "all"
-        else:
-            Warning('Given location or city: \
-                {} is not defined'.format(name))
-            self._city = ""
-
-    def set_sex(self, name: str):
-        """Set Corpus' instance gender by its name and check beforehand
-
-        Args:
-            name (str): Name of the gender the corpus should be filtered on
-        """
-        sex_list = self._feature_dict['Sex']
-        if name in sex_list:
-            self._gender = name
-        elif name == "all":
-            self._gender = "all"
-        else:
-            Warning('Gender not in accordance with corpus')
-            self._gender = ""
-
-    def set_age(self, name: str):
-        """Set Corpus' instance age by its name and check beforehand
-
-        Args:
-            name (str): Name of the age the corpus should be filtered on
-        """
-        age_list = self._feature_dict['Age']
-        if name in age_list:
-            self._age = name
-        elif name == "all":
-            self._age = "all"
-        else:
-            Warning('Age definition not in accordance with corpus')
-            self._age = ""
-
-    def set_education(self, name: str):
-        """Set Corpus' instance education by its name and check beforehand
-
-        Args:
-            name (str): Name of the education the corpus should be filtered on
-        """
-        education_list = self._feature_dict['Education']
-        if name in education_list:
-            self._education = name
-        elif name == "all":
-            self._education = "all"
-        else:
-            Warning('Education definition not in accordance with corpus')
-            self._education = ""
-
     def set_filter(self, city: str, gender: str,
                    age: str, education: str, phrase: str):
         self.set_city(city)
-        self.set_sex(gender)
+        self.set_gender(gender)
         self.set_age(age)
         self.set_education(education)
         self.set_search_phrase(phrase)
-
-        filter_name = "preseea_{}_{}_{}_{}_{}".format(city, gender, age, education, phrase)
-
-        return filter_name
 
     def get_filter(self):
         city = 'City: {}'.format(self._city)
@@ -640,77 +552,6 @@ class PRESEEA(Corpus):
 
         return countries
 
-    def get_all_cities(self) -> list:
-        """Return all the available cities in the Corpus
-
-        Returns:
-            list: List of strings with the city names
-        """
-
-        temporary_city_list = []
-        for country in self.get_corpus_countries():
-            temporary_city_list.append(self._feature_dict["Country"][country])
-
-        city_list = list(it.chain.from_iterable(temporary_city_list))
-
-        return city_list
-
-    def _check_city(self, sample: str) -> list:
-        """Check if samples are available for that feature
-
-        Args:
-            sample (str): Sample name, e.g. Colombia
-
-        Returns:
-            list: List of strings with available samples
-        """
-
-        # Compare demanded sample with the available feature list
-        try:
-            feature_list = self._feature_dict["Country"][sample]
-        except KeyError:
-            raise KeyError('{} not available in Corpus'.format(sample))
-
-        return feature_list
-
-    def get_cities(self, country: str) -> list:
-        """Return the cities available in the given country
-
-        Args:
-            country (str): Country within the Corpus
-
-        Returns:
-            list: List of strings with the available cities
-        """
-        city_list = self._check_city(country)
-
-        return city_list
-
-    def get_number_cities(self, country: str) -> int:
-        """Return the number of available cities per country
-
-        Args:
-            country ([type]): Name of the country
-        Returns:
-            int: Number of cities to be returned for country
-        """
-        city_list = self.get_cities(country)
-        number_cities = len(city_list)
-
-        return number_cities
-
-    def get_number_all_cities(self) -> int:
-        """Return the number of all available Corpus cities
-
-        Returns:
-            int: Number of available cities
-        """
-
-        city_list = self.get_all_cities()
-        number_cities = len(city_list)
-
-        return number_cities
-
     def create_report(self, city: str, phrase: str):
         """Create a .csv file as a report for the
            phrases and their corresponding analysis
@@ -723,11 +564,11 @@ class PRESEEA(Corpus):
                 in the corpus
         """
         # Define search issue with filters and phrase
-        filter_name = self.set_filter(city=city,
-                                      gender="all",  # "Hombre",
-                                      age="all",  # "Grupo 1",
-                                      education="all",  # "Bajo",
-                                      phrase=phrase)
+        self.set_filter(city=city,
+                        gender="all",  # "Hombre",
+                        age="all",  # "Grupo 1",
+                        education="all",  # "Bajo",
+                        phrase=phrase)
 
         # Get data via scrapy framework as API
         sample_list = self.retrieve_phrase_data()
@@ -736,6 +577,7 @@ class PRESEEA(Corpus):
         meta_data = self.analyse(sample_list)
 
         # Write data with stats into .csv file
+        filter_name = self.__str__()
         self.write_csv(file_name="report/{}.csv".format(filter_name),
                        data=sample_list,
                        meta=meta_data)
